@@ -29,7 +29,7 @@
   },
   "homepage": "https://github.com/seangenabe/stream-to-iterator#readme",
   "engines": {
-    "node": ">=6.5"
+    "node": ">=8.9"
   },
   "dependencies": {
     "p-defer": "1.0.0",
@@ -38,7 +38,7 @@
   "devDependencies": {
     "@types/node": "^9.4.1",
     "ava": "^1.0.0-beta.1",
-    "delay": "^2.0.0",
+    "into-stream": "^3.1.0",
     "typescript": "^2.7.1"
   },
   "ava": {
